IOT_project/requirements.txt
absl-py==0.14.0
astunparse==1.6.3
bleak==0.12.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
clang==5.0
cryptography==3.2
cycler==0.10.0
dm-tree==0.1.6
firebase-admin==5.0.3
flatbuffers==1.12
gast==0.4.0
google-api-core==2.1.0
google-api-python-client==2.25.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.20.1
idna==3.2
joblib==1.1.0
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
mpmath==1.2.1
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
paho-mqtt==1.5.1
pandas==1.3.3
pandas-stubs==1.2.0.29
Pillow==8.3.2
proto-plus==1.19.5
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyobjc-core==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-CoreBluetooth==7.3
pyobjc-framework-libdispatch==7.3
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
six==1.15.0
sklearn==0.0
sympy==1.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-model-optimization==0.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
